requirements.txt
torch>=2.1
pyyaml>=6.0
numpy>=1.24,<2.0  # Added version constraint to solve compatibility issues
scipy>=1.10,<1.14  # Added version constraint to solve compatibility issues
pillow>=10.0
transformers @ git+https://github.com/huggingface/transformers.git
peft>=0.6.0  # Updated version number
accelerate>=0.20.0  # Added missing package
scikit-learn>=1.3.0  # Added missing package
